Stop soft deleting tasks
This simplifies a lot of thing. Unfortunately the todoist API does not
expose completed or deleted tasks. Which adds quite a bit of complexity
on our side to handle it. So for now, stick to API behaviour. Later when
project is more mature, I would love to keep track of deleted and
performed tasks, but it's just too much work for now.